+++ http-clients/apache-client/src/main/java/software/amazon/awssdk/http/apache/internal/impl/ApacheHttpRequestFactory.java
@@ -24,7 +24,14 @@ import java.util.Optional;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpHeaders;
 import org.apache.http.client.config.RequestConfig;
+import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpEntityEnclosingRequestBase;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpHead;
+import org.apache.http.client.methods.HttpOptions;
+import org.apache.http.client.methods.HttpPatch;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.client.methods.HttpPut;
 import org.apache.http.client.methods.HttpRequestBase;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.http.HttpExecuteRequest;
@@ -109,18 +116,23 @@ public class ApacheHttpRequestFactory {
 
 
     private HttpRequestBase createApacheRequest(HttpExecuteRequest request, URI uri) {
-        SdkHttpMethod method = request.httpRequest().method();
-        switch (method) {
+        switch (request.httpRequest().method()) {
             case HEAD:
+                return new HttpHead(uri);
             case GET:
+                return new HttpGet(uri);
             case DELETE:
+                return new HttpDelete(uri);
             case OPTIONS:
+                return new HttpOptions(uri);
             case PATCH:
+                return wrapEntity(request, new HttpPatch(uri));
             case POST:
+                return wrapEntity(request, new HttpPost(uri));
             case PUT:
-                return wrapEntity(request, new HttpRequestImpl(method, uri));
+                return wrapEntity(request, new HttpPut(uri));
             default:
-                throw new RuntimeException("Unknown HTTP method name: " + method);
+                throw new RuntimeException("Unknown HTTP method name: " + request.httpRequest().method());
         }
     }
 
@@ -179,19 +191,5 @@ public class ApacheHttpRequestFactory {
         return !SdkHttpUtils.isUsingStandardPort(request.protocol(), request.port())
                 ? request.host() + ":" + request.port()
                 : request.host();
-    }
-
-    private static final class HttpRequestImpl extends HttpEntityEnclosingRequestBase {
-        private final SdkHttpMethod method;
-
-        private HttpRequestImpl(SdkHttpMethod method, URI uri) {
-            this.method = method;
-            setURI(uri);
-        }
-
-        @Override
-        public String getMethod() {
-            return this.method.toString();
-        }
     }
 }