+++ test/http-client-tests/src/main/java/software/amazon/awssdk/http/SdkHttpClientDefaultTestSuite.java
@@ -61,7 +61,7 @@ public abstract class SdkHttpClientDefaultTestSuite {
     @Test
     public void supportsResponseCode200Head() throws Exception {
         // HEAD is special due to closing of the connection immediately and streams are null
-        testForResponseCode(HttpURLConnection.HTTP_FORBIDDEN, SdkHttpMethod.HEAD, false);
+        testForResponseCode(HttpURLConnection.HTTP_FORBIDDEN, SdkHttpMethod.HEAD);
     }
 
     @Test
@@ -76,7 +76,7 @@ public abstract class SdkHttpClientDefaultTestSuite {
 
     @Test
     public void supportsResponseCode403Head() throws Exception {
-        testForResponseCode(HttpURLConnection.HTTP_FORBIDDEN, SdkHttpMethod.HEAD, false);
+        testForResponseCode(HttpURLConnection.HTTP_FORBIDDEN, SdkHttpMethod.HEAD);
     }
 
     @Test
@@ -98,64 +98,22 @@ public abstract class SdkHttpClientDefaultTestSuite {
     public void validatesHttpsCertificateIssuer() {
         SdkHttpClient client = createSdkHttpClient();
 
-        SdkHttpFullRequest request = mockSdkRequest("https://localhost:" + mockServer.httpsPort(), SdkHttpMethod.POST, true);
+        SdkHttpFullRequest request = mockSdkRequest("https://localhost:" + mockServer.httpsPort(), SdkHttpMethod.POST);
 
         assertThatThrownBy(client.prepareRequest(HttpExecuteRequest.builder().request(request).build())::call)
             .isInstanceOf(SSLHandshakeException.class);
     }
 
-    @Test
-    public void supportsRequestBodyOnGetRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.GET, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnPostRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.POST, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnPutRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.PUT, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnDeleteRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.DELETE, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnHeadRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.HEAD, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnPatchRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.PATCH, true);
-    }
-
-    @Test
-    public void supportsRequestBodyOnOptionsRequest() throws Exception {
-        testForResponseCode(200, SdkHttpMethod.OPTIONS, true);
-    }
-
     private void testForResponseCode(int returnCode) throws Exception {
-        testForResponseCode(returnCode, SdkHttpMethod.POST, true);
+        testForResponseCode(returnCode, SdkHttpMethod.POST);
     }
 
-    protected void testForResponseCode(int returnCode, SdkHttpMethod method, boolean includeBody) throws Exception {
-        testForResponseCode(returnCode, method, method, includeBody);
-    }
-
-    protected void testForResponseCode(int returnCode,
-                                       SdkHttpMethod method,
-                                       SdkHttpMethod expectedMethod,
-                                       boolean includeBody) throws Exception {
+    private void testForResponseCode(int returnCode, SdkHttpMethod method) throws Exception {
         SdkHttpClient client = createSdkHttpClient();
 
         stubForMockRequest(returnCode);
 
-        SdkHttpFullRequest req = mockSdkRequest("http://localhost:" + mockServer.port(), method, includeBody);
+        SdkHttpFullRequest req = mockSdkRequest("http://localhost:" + mockServer.port(), method);
         HttpExecuteResponse rsp = client.prepareRequest(HttpExecuteRequest.builder()
                                                                           .request(req)
                                                                           .contentStreamProvider(req.contentStreamProvider()
@@ -163,14 +121,14 @@ public abstract class SdkHttpClientDefaultTestSuite {
                                                                           .build())
                                         .call();
 
-        validateResponse(rsp, returnCode, expectedMethod, includeBody);
+        validateResponse(rsp, returnCode, method);
     }
 
     protected void testForResponseCodeUsingHttps(SdkHttpClient client, int returnCode) throws Exception {
         SdkHttpMethod sdkHttpMethod = SdkHttpMethod.POST;
         stubForMockRequest(returnCode);
 
-        SdkHttpFullRequest req = mockSdkRequest("https://localhost:" + mockServer.httpsPort(), sdkHttpMethod, true);
+        SdkHttpFullRequest req = mockSdkRequest("https://localhost:" + mockServer.httpsPort(), sdkHttpMethod);
         HttpExecuteResponse rsp = client.prepareRequest(HttpExecuteRequest.builder()
                                                                           .request(req)
                                                                           .contentStreamProvider(req.contentStreamProvider()
@@ -178,7 +136,7 @@ public abstract class SdkHttpClientDefaultTestSuite {
                                                                           .build())
                                         .call();
 
-        validateResponse(rsp, returnCode, sdkHttpMethod, true);
+        validateResponse(rsp, returnCode, sdkHttpMethod);
     }
 
     private void stubForMockRequest(int returnCode) {
@@ -193,20 +151,17 @@ public abstract class SdkHttpClientDefaultTestSuite {
         mockServer.stubFor(any(urlPathEqualTo("/")).willReturn(responseBuilder));
     }
 
-    private void validateResponse(HttpExecuteResponse response,
-                                  int returnCode,
-                                  SdkHttpMethod method,
-                                  boolean expectBody) throws IOException {
+    private void validateResponse(HttpExecuteResponse response, int returnCode, SdkHttpMethod method) throws IOException {
         RequestMethod requestMethod = RequestMethod.fromString(method.name());
 
         RequestPatternBuilder patternBuilder = RequestPatternBuilder.newRequestPattern(requestMethod, urlMatching("/"))
                                                                     .withHeader("Host", containing("localhost"))
                                                                     .withHeader("User-Agent", equalTo("hello-world!"));
 
-        if (expectBody) {
-            patternBuilder.withRequestBody(equalTo("Body"));
-        } else {
+        if (method == SdkHttpMethod.HEAD) {
             patternBuilder.withRequestBody(absent());
+        } else {
+            patternBuilder.withRequestBody(equalTo("Body"));
         }
 
         mockServer.verify(1, patternBuilder);
@@ -222,14 +177,14 @@ public abstract class SdkHttpClientDefaultTestSuite {
         mockServer.resetMappings();
     }
 
-    private SdkHttpFullRequest mockSdkRequest(String uriString, SdkHttpMethod method, boolean includeBody) {
+    private SdkHttpFullRequest mockSdkRequest(String uriString, SdkHttpMethod method) {
         URI uri = URI.create(uriString);
         SdkHttpFullRequest.Builder requestBuilder = SdkHttpFullRequest.builder()
                                                                       .uri(uri)
                                                                       .method(method)
                                                                       .putHeader("Host", uri.getHost())
                                                                       .putHeader("User-Agent", "hello-world!");
-        if (includeBody) {
+        if (method != SdkHttpMethod.HEAD) {
             byte[] content = "Body".getBytes(StandardCharsets.UTF_8);
             requestBuilder.putHeader("Content-Length", Integer.toString(content.length));
             requestBuilder.contentStreamProvider(() -> new ByteArrayInputStream(content));